Fix link optimization
In general linker will omit object files which doesn't contain any
of the referred symbols. This breaks logic of registering module. The
previous solution uses function "isActive" at the point of injection
an this enforces linking of library. But this solution prevents
inlining of check and can slow down resulting application. The new
solution calls "registerModule" only when fault injection point is
active so it slow down only when injection is triggered.