Add support for injection point versioning
The older version is 0. The newer version is 1. The data from the next
versions should be placed to union in structure.

To distinguish versions the very first member of structure is version
value. The 0 version has pointer in the first member so its value
always greater that 256 and allows to detect this version. As the
result maximum version of structure is 255.